cloudrun_service/requirements.txt
functions-framework==3.7.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.25.0
apache-beam[gcp]==2.57.0
python-dateutil==2.9.0.post0
